Add support for RGB bulbs.
These can be set to 24 bit RGB values, but are only capable of reporting 8
bit RGB values (ie, 3 bits of red, 3 bits of green, 2 bits of blue). We fudge
these back to a plausible value. However, in addition, brightness is an
orthogonal value - I need to go and study some more colour theory to figure
out how to map this correctly.